Optimized the organizational structure of the code
What did this commit do:
Adjusted the organizational structure of the code and split the functional modules
into independent crates in the workspace

Why do this:
1. Make it more convenient to maintain and test individual functional modules
2. Make it more convenient to quickly build new binary based on basic functional components
