+++ Makefile
@@ -1,30 +1,30 @@
 .PHONY: kbs
 kbs:
-	cargo build
+	cargo build --bin kbs
 
 .PHONY: kbs-coco-as
 kbs-coco-as:
-	cargo build --no-default-features --features coco-as-builtin,rustls
+	cargo build --bin kbs --no-default-features --features coco-as-builtin,rustls
 
 .PHONY: kbs-coco-as-grpc
 kbs-coco-as-grpc:
-	cargo build --no-default-features --features coco-as-grpc,rustls
+	cargo build --bin kbs --no-default-features --features coco-as-grpc,rustls
 
 .PHONY: kbs-coco-as-no-verifier
 kbs-coco-as-no-verifier:
-	cargo build --no-default-features --features coco-as-builtin-no-verifier,rustls
+	cargo build --bin kbs --no-default-features --features coco-as-builtin-no-verifier,rustls
 
 .PHONY: kbs-coco-as-openssl
 kbs-coco-as-openssl:
-	cargo build --no-default-features --features coco-as-builtin,openssl
+	cargo build --bin kbs --no-default-features --features coco-as-builtin,openssl
 
 .PHONY: kbs-coco-as-grpc-openssl
 kbs-coco-as-grpc-openssl:
-	cargo build --no-default-features --features coco-as-grpc,openssl
+	cargo build --bin kbs --no-default-features --features coco-as-grpc,openssl
 
 .PHONY: kbs-coco-as-no-verifier-openssl
 kbs-coco-as-no-verifier-openssl:
-	cargo build --no-default-features --features coco-as-builtin-no-verifier,openssl
+	cargo build --bin kbs --no-default-features --features coco-as-builtin-no-verifier,openssl
 
 .PHONY: check
 check:
