+++ attestation/src/coco/mod.rs
@@ -1,0 +1,5 @@
+#[cfg(feature = "coco-as-grpc")]
+pub mod grpc;
+
+#[cfg(any(feature = "coco-as-builtin", feature = "coco-as-builtin-no-verifier"))]
+pub mod builtin;
