+++ docker/Dockerfile.coco-as-grpc
@@ -6,7 +6,7 @@ COPY . .
 RUN apt-get update && apt install -y protobuf-compiler
 
 # Build and Install KBS
-RUN cargo install --path src/kbs --no-default-features --features coco-as-grpc,rustls
+RUN cargo install --path bin/kbs --no-default-features --features coco-as-grpc,rustls
 
 
 FROM ubuntu:22.04
