+++ session/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "session"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+documentation.workspace = true
+edition.workspace = true
+
+[dependencies]
+actix-web.workspace = true
+aes-gcm = "0.10.1"
+anyhow.workspace = true
+as-types.workspace = true
+base64.workspace = true
+internal.workspace = true
+kbs-types.workspace = true
+rand = "0.8.5"
+rsa = "0.7.2"
+semver.workspace = true
+serde_json.workspace = true
+tokio.workspace = true
+uuid = { version = "1.2.2", features = ["serde", "v4"] }
