--- src/api/Cargo.toml
@@ -1,58 +1,0 @@
-[package]
-name = "api-server"
-version.workspace = true
-authors.workspace = true
-description.workspace = true
-documentation.workspace = true
-edition.workspace = true
-
-[features]
-default = ["coco-as-builtin", "rustls"]
-coco-as = []
-coco-as-builtin = ["coco-as", "attestation-service/default"]
-coco-as-builtin-no-verifier = ["coco-as", "attestation-service/rvps-native"]
-coco-as-grpc = ["coco-as", "tonic", "tonic-build", "prost"]
-rustls = ["actix-web/rustls", "dep:rustls", "dep:rustls-pemfile"]
-openssl = ["actix-web/openssl", "dep:openssl"]
-
-[dependencies]
-actix-web = "4"
-actix-web-httpauth = "0.8.0"
-aes-gcm = "0.10.1"
-anyhow.workspace = true
-async-trait.workspace = true
-# TODO: update to as-types rev=v0.6.0 once released
-as-types = { git = "https://github.com/confidential-containers/attestation-service.git", rev = "f119a5d"}
-# TODO: update to attestation-service rev=v0.6.0 once released
-attestation-service = { git = "https://github.com/confidential-containers/attestation-service.git", default-features = false, rev = "f119a5d", optional = true}
-base64.workspace = true
-cfg-if = "1.0.0"
-elliptic-curve = { version = "0.13.4", features = ["arithmetic", "pem"] }
-env_logger.workspace = true
-jwt-simple = "0.11.4"
-kbs-types = "0.3"
-lazy_static = "1.4.0"
-log.workspace = true
-prost = { version = "0.11", optional = true }
-rand = "0.8.5"
-rcgen = "0.10.0"
-rsa = "0.7.2"
-rustls = { version = "0.20.8", optional = true }
-rustls-pemfile = { version = "1.0.2", optional = true }
-semver = "1.0.16"
-serde = { version = "1.0", features = ["derive"] }
-serde_json.workspace = true
-strum = "0.24.1"
-strum_macros = "0.24.1"
-tokio.workspace = true
-tonic = { version = "0.8", optional = true }
-uuid = { version = "1.2.2", features = ["serde", "v4"] }
-openssl = { version = "0.10.46", optional = true }
-p256 = { version = "0.13.2", features = [ "ecdsa" ] }
-
-[dev-dependencies]
-tempfile.workspace = true
-
-[build-dependencies]
-anyhow = "1"
-tonic-build ={ version = "0.8", optional = true }
