--- src/api/src/attestation/coco/mod.rs
@@ -1,5 +1,0 @@
-#[cfg(feature = "coco-as-grpc")]
-pub(crate) mod grpc;
-
-#[cfg(any(feature = "coco-as-builtin", feature = "coco-as-builtin-no-verifier"))]
-pub(crate) mod builtin;
