--- src/kbs/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "kbs"
-version.workspace = true
-authors.workspace = true
-description.workspace = true
-documentation.workspace = true
-edition.workspace = true
-
-[features]
-default = ["coco-as-builtin", "rustls"]
-coco-as-builtin = ["api-server/coco-as-builtin"]
-coco-as-builtin-no-verifier = ["api-server/coco-as-builtin-no-verifier"]
-coco-as-grpc = ["api-server/coco-as-grpc"]
-rustls = ["api-server/rustls"]
-openssl = ["api-server/openssl"]
-
-[dependencies]
-anyhow.workspace = true
-api-server.workspace = true
-clap = { version = "4.0.29", features = ["derive"] }
-env_logger.workspace = true
-log.workspace = true
-tokio.workspace = true
