+++ token/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "token"
+version.workspace = true
+authors.workspace = true
+description.workspace = true
+documentation.workspace = true
+edition.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+base64.workspace = true
+elliptic-curve = { version = "0.13.4", features = ["arithmetic", "pem"] }
+jwt-simple.workspace = true
+p256 = { version = "0.13.2", features = [ "ecdsa" ] }
+rcgen = "0.10.0"
+serde.workspace = true
+serde_json.workspace = true
+strum.workspace = true
+strum_macros.workspace = true
+tokio.workspace = true
